refactor(auth): 파일명 케밥 케이스로 변경 및 파일 위치 변경
