+++ src/app/auth/signup/email/page.tsx
@@ -1,0 +1,12 @@
+' use client';
+
+import EmailSignup from '@/app/features/auth/signup/email/components/email-signup';
+
+const EmailSignupPage = () => {
+  return (
+    <div>
+      <EmailSignup />
+    </div>
+  );
+};
+export default EmailSignupPage;
