+++ src/app/features/auth/signup/email/components/email-signup.tsx
@@ -1,0 +1,11 @@
+const EmailSignup = () => {
+  return (
+    <div className='flex items-center justify-center'>
+      <div>
+        <p>회원가입</p>
+      </div>
+    </div>
+  );
+};
+
+export default EmailSignup;
